Moved IPC sink test to correct directory.
