+++ src/io/ipc/write/file_async.rs
@@ -247,54 +247,3 @@ where
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::FileSink;
-    use crate::{
-        array::{Array, Float32Array, Int32Array},
-        chunk::Chunk,
-        datatypes::{DataType, Field, Schema},
-        io::ipc::read::file_async::{read_file_metadata_async, FileStream},
-    };
-    use futures::{io::Cursor, SinkExt, TryStreamExt};
-    use std::sync::Arc;
-
-    // Verify round trip data integrity when using async read + write.
-    #[test]
-    fn test_file_async_roundtrip() {
-        futures::executor::block_on(async move {
-            let mut data = vec![];
-            for i in 0..5 {
-                let a1 = Int32Array::from(&[Some(i), None, Some(i + 1)]);
-                let a2 = Float32Array::from(&[None, Some(i as f32), None]);
-                let chunk = Chunk::new(vec![
-                    Arc::new(a1) as Arc<dyn Array>,
-                    Arc::new(a2) as Arc<dyn Array>,
-                ]);
-                data.push(chunk);
-            }
-            let schema = Schema::from(vec![
-                Field::new("a1", DataType::Int32, true),
-                Field::new("a2", DataType::Float32, true),
-            ]);
-
-            let mut buffer = Cursor::new(Vec::new());
-            let mut sink = FileSink::new(&mut buffer, schema.clone(), None, Default::default());
-            for chunk in &data {
-                sink.feed(chunk.clone()).await.unwrap();
-            }
-            sink.close().await.unwrap();
-            drop(sink);
-
-            buffer.set_position(0);
-            let metadata = read_file_metadata_async(&mut buffer).await.unwrap();
-            assert_eq!(schema, metadata.schema);
-            let stream = FileStream::new(buffer, metadata, None);
-            let out = stream.try_collect::<Vec<_>>().await.unwrap();
-            for i in 0..5 {
-                assert_eq!(data[i], out[i]);
-            }
-        })
-    }
-}
