+++ tests/it/io/ipc/mod.rs
@@ -9,3 +9,6 @@ mod write_async;
 
 #[cfg(feature = "io_ipc_read_async")]
 mod read_stream_async;
+
+#[cfg(feature = "io_ipc_read_async")]
+mod read_file_async;
